pyproject.toml
[project]
name = "collam"
version = "0.1.0"
description = "collam"
readme = "README.md"
requires-python = ">=3.9, <4.0"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
    { name = "Console", module_name = "nonebot.adapters.console" },
]
plugins = []
plugin_dirs = ["src/plugins"]
builtin_plugins = ["echo"]

[tool.poetry.dependencies]
# ============= NoneBot 核心及适配器 =============
nonebot2 = { extras = ["fastapi"], version = "^2.4.1" }
nonebot-adapter-onebot = "^2.4.6"
nonebot-adapter-console = "^0.6.0"

# ============= HTTP/网络库 =============
httpx = "^0.28.1"
httpcore = "^1.0.7"
websockets = "^15.0"
yarl = "^1.18.3"     # URL 解析库
anyio = "^4.8.0"     # 异步兼容层

# ============= 数据处理与序列化 =============
pydantic = "^2.10.6"
pydantic-core = "^2.27.2"
msgpack = "^1.1.0"
annotated-types = "^0.7.0"

# ============= 实用工具库 =============
python-dotenv = "^1.0.1" # 环境变量加载
loguru = "^0.7.3"        # 日志工具
rich = "^13.9.4"         # 终端美化
tqdm = "^4.67.1"         # 进度条
platformdirs = "^4.3.6"  # 跨平台目录管理

# ============= 开发与调试 =============
ipython = "^8.0.0"   # 推荐添加的交互式调试工具
pygments = "^2.19.1" # 语法高亮
textual = "^2.1.2"   # 终端 UI 框架

# ============= 安全相关 =============
certifi = "2025.1.31"   # CA 证书（必须严格锁定）
cryptography = "43.0.3" # 加密库（推荐锁定）

# ============= 平台特定依赖 =============
win32-setctime = { version = "^1.2.0", platform = "win32" }

# ============= 底层依赖（建议锁定） =============
cffi = "1.17.1"
pycparser = "2.22"
idna = "3.10"      # 域名处理

# ============= 其他依赖 =============
colorama = "^0.4.6"
distro = "1.9.0"
exceptiongroup = "1.2.2"
h11 = "0.14.0"
jiter = "0.8.2"
linkify-it-py = "2.0.3"
markdown-it-py = "3.0.0"
mdit-py-plugins = "0.4.2"
mdurl = "0.1.2"
multidict = "6.1.0"
nonechat = "0.3.0"           # 注意：主版本为 0 需谨慎
propcache = "0.3.0"
pygtrie = "2.5.0"
sniffio = "1.3.1"
tomli = "2.2.1"
typing-extensions = "4.12.2"
uc-micro-py = "1.0.3"
